Update stellar_spectrum.py
Added a warning to the user (in the form of a terminal output) if the pySME output is unusually short. This indicates that a line list is used which does not include spectral lines in the user-specified wavelength range. StarRotator will still produce all the files but they will be basically just 1-3 datapoints inside of no use for any correction.

This case has happened both to me and a student using the pySME module (with the demo linelist, but not on the sodium order).